fix MSVC

+++ main.c
@@ -340,7 +340,7 @@ int main(int argc, char **argv) {
 				const Line *line = editor.lines + row;
 				render_text_sized(
 					renderer, &font, line->chars, line->size,
-					vec2f(0.0f, row * FONT_CHAR_HEIGHT * FONT_SCALE),
+					vec2f(0.0f, (float)row * FONT_CHAR_HEIGHT * FONT_SCALE),
 					0xFFFFFFFF, FONT_SCALE);
 			}
 			render_cursor(renderer, &font);
